Some SkPath cleanups
- use enum to store filltype
- share function to toggle inverse-filltype
- rm unused methods

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1078596
Commit-Queue: Mike Reed <mike@reedtribe.org>

+++ src/core/SkPath_serial.cpp
@@ -91,7 +91,7 @@ size_t SkPath::writeToMemoryAsRRect(void* storage) const {
         return sizeNeeded;
     }
 
-    int32_t packed = (fFillType << kFillType_SerializationShift) |
+    int32_t packed = (static_cast<int>(fFillType) << kFillType_SerializationShift) |
                      ((int)firstDir << kDirection_SerializationShift) |
                      (SerializationType::kRRect << kType_SerializationShift) |
                      kCurrent_Version;
@@ -112,7 +112,7 @@ size_t SkPath::writeToMemory(void* storage) const {
         return bytes;
     }
 
-    int32_t packed = (fFillType << kFillType_SerializationShift) |
+    int32_t packed = (static_cast<int>(fFillType) << kFillType_SerializationShift) |
                      (SerializationType::kGeneral << kType_SerializationShift) |
                      kCurrent_Version;
